dl manager to main
